Complete rewrite of the data loading pipeline.
This commit rewrites the data loading pipeline to use a somewhat larger number
of more concrete, and much simpler, classes:

    * Subject    : the individual imaged in the scan
    * Aquisition : the metadata for a scan
    * Volume     : the possibly labelled data for a scan
    * Dataset    : a set of aquisitions for training / validationn / test
    * Batch      : a set of patches and labels

The expected pattern is that most consumers of data will access it either by a
dataset object, or by a batch constructed from a dataset.

The dataset itself represents a partition of set of aquisitions, into a
training, validation and test set, randomised so as to ensure that data from a
single subject is not included in more than one component of the partition.

Each aquisition acts as both a descriptor for a scan, as well as a mechanism by
which to load the associated data on demand.  The separation of descriptor and
data allows the full data set to be queried without the requirement that all
data be loaded.  Because aquisitions are resposible for loading data, they will
almost necessarily be specific to each type of data, and probably specific to
the layout of each data source.

For the current project, an OASIS dataset specific aquisition class has been
defined for that purpose.

In order to support training, volumes associate the underlying image data
with class labels and a mask; for unlabelled data, the label component
will simply be empty.  The intent is that the volume acts as a fairly simple
generic wrapper for image data and its associated labels.