Added classes to define, model, optimise and train a neural network.
In this commit we introduce classes to represent the structure of the network,
(network.Layer and network.Architecture) as well as a class which maintains the
state of the learnable parameters (network.Model).  In addition to these, the
network module provides functions to train a network for one or multiple
epochs.

The above classes require concrete implementations of the network layers, the
cost function and the per batch optimisation function to compute the new value
of the parameters given the current value and the cost.  Several common layer
types are implemented in the layers module, while the standard cross entropy
cost function, along with a very simple gradient descent optimisation step are
provided in the optimisation module.