Added tests for the data module and corrected data.Batch.normalised_bounds.
The tests made clear that actually the normalised bounds which were being
constructed were much larger than required.  These were previously equal to
the smallest minimum to the largest maximum, but actually we just want the
largest difference between a maximum and a minimum, as that is the smallest
bounds which will contain the unmasked region in each volume.

Some of the new tests for the data module are still just stubs, but most
things are now properly tested, which is nice.