Refactored experiments, labels and data modules to factor out label transforms.
This commit primarily serves to factor out the logic of reshaping output and ground
truth labels into the appropriate formats, into a policy class (in labels) which is
called from the new Accessor class (in data) used by the experiments.

As part of these changes, the epxeriments were simplified to remove the need for
subclasses, which primarily dealt with these transformations.  As part of that, the
event handlers were filled out with basic functionality.