Factored output handling out from Experiment into a separate class.
The experiment class is where all of the other components are brought
together, so in a sense it is a very broad class.  To also add to it new
functionality of its own makes it over-complex.  I have moved the functions
which handle the on_batch and on_epoch events during training into a new
class whose sole purpose is recording training output, and modified the
experiment constructor to accept an instance of this output class as a
parameter.  The actual name and form of this class is likely to change
very substantially when I get to the point of adding actual experiments,
but the most important thing here is that the functionality is now outside
of the experiment class itself.