Corrected an error in the categorical cross entropy calculation.
This was previously computed as a mean over the computed probabilities
weighted by the corresponding true probability, but on reflection it is
obvious that this is not the mean cross entropy, which would be the
mean of the sums over computed probabilities weighted by the true
probability.  The difference here would of course only be a constant
linear factor, specifically the old version would compute a cost which
was C times smaller than the correct value, where C is the number of
classes in the distribution.  Nonetheless this reduced cost might cause
learning to be much slower than it should be.