Small corrections to fix errors in the layers module.
The layers module contained a number of instances in which an assignment
operator was used instead of an equality (presumably one initial error
which was then copied and pasted multiple times).  It also contained a
number of places where a module was invoked using a name different to the
one by which it was imported.  Finally there were places where the wrong
a function which computed its result in multiple steps and something other
than the final step was incorrectly returned.  All of these have now been
corrected.