Tests for the data module and corrected an issue with the offset representation
This commit adds test for patch offsets and sampling in the Batch class, as
well as a test of the somewhat non-trivial constructor.  These demonstrated
that there was an issue with the representation of the offsets, which had the
per volume axis after the minimum vs maximum axis, meaning that views would
be very inefficient (as these are currently always per-volume).  This was
corrected with a permutation.

The tests also showed that the interfaces to some functions could be simplified
to note require a volume when only certain properties were truly required.

Tests were also added for all of the data parameter replication methods.