Provided the results module with access to the shape of the patch grid.
The label module has functions to reconstruct a dense labelling of the
volume from a set of dense patches, or sparse per patch labels, provided
that the shape of the patches and the grid from which they were sampled
is known.  Unfortunately the only place which previously knew about the
grid was deep down in the Batch class, while the function invoking the
labels module was in the experiments.  This commit exposes the patch
grid shape for a batch, which is now returned by the data accessor along
with the input and the labels, and this is then passed along to the
experiment results event handlers by the network trainer... whewh.  That
was more complicated than I would have liked.  Sadly however, because the
shape of the grid can change between batches, there is really no other
way to do this.

This commit actually captures a few other small fixes to the modules
along this chain, as I don't really have time to split these into a
separate commit right now.