Rewrite the experiment output class and renamed this Results.
The results module now has primary responsibility for collating results
and writing these to disk.  It receives results during training after
its on_batch and on_epoch functions are registered as handlers for those
events with the network trainer.

The set of metrics for the results has been expanded (we now have DICE
scores) as has the set of images which can be produced for these (we
now have confusion matrix images for each class).  The actual metric
and image functions have been moved into their own classes, Images and
Metrics respectively on the expectation that this set of functions is
likely to grow (and given that the Results class was already fairly
large, and responsible for quite a lot).