Added a basic real data test and corrected Theano/CUDA pipeline errors.
As this is the first full end to end test of this version of the code,
a number of issues surfaced and were dealt with.  These primarily concerned
the dimensions and dtypes of Theano variables, but also some small logical
errors in the way that different stages of the pipeline were connected.
The pipeline now runs, though with too little output to be very comfortable,
but addressing that is a task for the next commit.