Added log out put to most modules used during training.
Training is a complex, and potentially very long running process,
so having some feedback about the current status is important.
This commit adds a class for logging with some basic semantics,
and adds log output to most of the classes which are responsible
for some part of the training process.  This functionality may
be extended in future with support for logging the time taken
for operations, and perhaps the system resource usage at various
points, but this has not yet been implemented.