Added support for splitting volume patches across batches, and randomising.
Previously the data module assumed that all patches from a volume were
processed in a single batch (though a btach could still contain patches
from multiple volumes).  This approach fails if the network size is too
great to process a whole volume in a single pass.  This commit allows the
patches for a volume to be divided across batches, and also randomised,
which becomes important if only a few patches are to be processed at once.

Note that ContiguousPatchSet (representing a set of patches densely covering
a region of the volume in non-random order) is not yet tested, though the
RandomPatchSet is.  The tests for the former class will be the subject of
the next commit (committing now to allow the current changes to be pushed to
another machine).