Modified the PatchSet to return a flat list of patches; completed tests.
In the course of completing the tests for the ContiguousPatchSet, it
became obvious that the existing pattern, of returning patches grouped
by volume does not work as one might expect where the number of patches
differs between the volumes.  This was not previously a problem, as all
volumes contributed the same number of patches by design, but with the
new ability to split volumes between multiple batches this is no longer
the case.

This will require changes to the way in which volumes are reconstructed,
but then the fact that volumes may be split between multiple batches had
already invalidated the existing method anyway.  The update required to
support these two changes have so far not been made, but will come in the
next few commits.

This commit also finishes the tests for the ContiguousPatchSet class and
corrects a number of errors in its implementation.