Moved model training from the network to a largely rewritten optimisation module.
This change was motivated by two concerns: (i) it should be possible to load a
model, and run it on an input volume, without the need for an optimiser or even
a cost function; (ii) both the network and training are about to become more
complicated (by the addition of serialisation to the former, by the addition of
different learning rate schedulers to the latter).  Both of these argue for
moving the training (and validation) functions from the network and into the
optimiser, effectively inverting the relationship which formerly existed: where
previously the model took an optimiser, and used it to construct the parameter
updates in the training pipeline, whereas in the new form the optimiser takes a
model and uses it to generate the predicted labels from which the costs and
parameter updates are then generated (and the latter applied to the model by
the optimiser).

As part of this redesign, the learning rate was split out of the optimiser itself
to facilitate wriing a more sophisticated scheduler; this initial version includes
an implementation of RMSProp, but should allow Adadelta and other advanced learning
rate schedulers as well.