Added an iterator over the batches in a dataset constructing the required patch sets.
The PatchSet class which represents the image and label patches for a single batch is
stateless, hence iteration over the batches for an epoch requires the caller to keep
track of the current state with which to correctly select and construct the next patch
set.  This commit wraps this behind an iterable Accessor type which takes a set of
aquisitions, sample parameters, and functions to normalise image data and convert
label.  This commit wraps this behind an iterable Accessor type.  Accessor instances
wrap a set of aquisitions, sample parameters, and functions to normalise image data and
convert labels and expose iteration over them behind a simple iterator interface.  This
allows the data to be accessed in a context in which these additional parameters
could not be determined (e.g. deep within the optimiser).

Introducing this class required two significant changes to the way that PatchSet classes
work:

  - all patch sets now take a target size for the sampled region of each volume; this
    allows the accessor to calculate the number of batches in an epoch without having
    to know the internal details of each volume.  Previously only the sequential patch
    set took such an argument, but this is now used generally, i.e. for random patch
    sets also.

  - patch offsets are no longer structures per volume, but instead are represented by a
    flat 2D array, with one row for each patch, but with an additional column giving the
    index of the volume from which the patch was extracted.  This allows the offsets for
    adjacent batches to be concatenated, without the need to keep a separate record of
    the volume indices (by for example recording all offsets from the beginning of the
    epoch).  This in turn will allow output for a volume to be accumulated across multiple
    batches and then reconstructed once all patches have been returned via the accessor.