Added an initial version of the optimisation monitors and the results classes.
These classes are responsible for aggregating output from the classifier, and
then generating both images and metrics to allow for subsequent evaluation.
At this point these classes are reasonably complete, but not yet integrated into
the experiemnt class, which is still using the older interface to pretty much
everything and will need to be rewritten.