Modified the label conversion classes to work on single volumes.
The functions in this module previously assumed that their input
was structured by volume (and then patch) but this is no longer
the case, now that the PatchSet returns a flat list.  As these
functions make heavy use of permutations on the order of axes,
this change was not so trivial as one might have expected, but it
is done now, and the tests all pass (after much debugging ;-).