Correction to the accumulated sum initialiser in the RMSProp learning rate.
This previously assumed that the value of all parameters was tensor, and
would thus fail in the case of a scalar; this now correctly checks for the
type of the initial value of the parameters whose gradients determine the
learning rate, and sets the type of accumulated sum accordingly (i.e. to
an array or a scalar 0).