Numerous fixes in network layers for axis order, dimensionality and overflow.
The layer module had a number of issues which only became apparent on real
data.  In numerous cases the dimensions over which an operation needed to
broadcast were not all to the left of the dimensions to which the operation
actually applied; this has been corrected by shuffling the target axes to
the end, applying the operation and then shuffling everything back.

A related issue affected the output of the network, where this is intended
to be a distribution over some set of classes at each point in a patch.  In
this case the axes should have the form P x X_n x ... x X_1 x C where P is
the index of the patch, X_n,...,X_1 are the dimensions of the patch and C
is the number classes in the distribution.  Given the way that softmax works
however, we actually had an output in the form P x C x X_n x ... x X_1,
with effectively one channel per class.  To resolve this we add a new layer
type which simply permutes its input to move the 2nd channel axis to the
end.

The last issue resolved in this commit affects the softmax function
itself.  While the initial implementation was correct, it had a tendency
to result in both over and underflow for many inputs, thereby littering
the output with NaN values.  I have reimplemented this in log form:

e^{P_i} / sum_j( e^{P_j} )  =  e^{log( e^{P_i} / sum_j( e^{P_j} ) )}
                            =  e^{log( e^{P_i} ) - log( sum_j( e^{P_j} ) )}
                            =  e^{log( e^{P_i} ) - sum_j( log( e^{P_j} ) )}
                            =  e^{P_i - sum_j( P_j )}

where the exponent can be computed first with no chance of overflow (given
that sum_j( p_j ) = 1) and only minimal chance of underflow (where P_i is
very small relative to 1.0).