Rewrite the experiments class; changed the results to accept a volume.
The rewrite basic served to do the following:

  - split out the components which need to be defined for each concrete
    experiment into an experiment definition class with a well defined
    interface

  - reimplemented most experiment properties to support lazy loading so
    that this can be done on execution of the experiment, and not on the
    construction of the experiment object; provided that these properties
    are not accessed outside of the class, this ensures that the output
    to the log happens in a sensible order.  If they are accessed prior
    to running the experiment, then the component objects wil of course
    be constructed earlier, and the order in the log will reflect this,
    but that is as it should be.

  - the run method now uses the new optimiser interface (optimise until
    converged) as opposed to the old network train method.

There are now subclasses of this for architectures which result in sparse
labels (i.e. one per patch) and those which give dense outputs.  These
serve to encapsulate the additional logic required to convert the GT
labels into the form erquired for the cost function, and conversely to
reconstruct the output of the network into a labelled volume for the
results class to process.

The interface with the results class has also changed: the monitor which
collects patches as they come out the validation step until it has a
whole volume, now reconstructs that volume itself before passing this to
the results.  This avoids needing to inform the results class of the logic
required to perform the reconstruction based on the format of the patches.

The other change here was a rename of the module from experiment to
experiments.  This was probably badly timed, but it stopped pylint from
nagging quite so much about shadowed variable names...