Modified results to save offsets, calculate statistics and add image types.
There are several changes to the results module here.  The changes to the
core SegmentationResults class itself involve:

  - persisting the offset for each computed distribution relative to the
    source volume

  - exposing statistics on each metric (i.e. the mean, media, minimum and
    maximum values) through the results object itself

The first of these features required changes to the experiment and sample
modules as well in order to pass through the requiredd offset information.

The statistics now have a type to represent them, and this object records
not only the value of the statistic but also the index of a volume which
is representative of it (i.e. the volume whose value for the metric is
closest to that statistic).  This is important for selecting sample images.

This commit also adds new types of sample images to the images class, to
support displaying multiple classes on a single image, along with the
true and false positives for each of these classes.  This is important
for getting a sense of the output of the classifier (i.e. seeing all the
classes at once) while also being able to compare this to the reference
(i.e. for each voxel labelled with a given class, indicating whether it
was labelled correctly or not).  This does not show shich class it should
have been labelled as of course, as there is only so much that can be
shown in one image, but it is good enough for an overview.  The binary
class images can be used for more detailed analysis.