Added a file extension for saved arrays and corrected corrected a loading bug.
The load_saved_array method on the archive did not previously return the array
which it loaded, but rather only the file name.  Not sure when this bug was
introduced, but it's gone now.

Neither the save nor load specified the file extension to use when saving arrays,
but while the underlying numpy save operation defaults to adding one, the load
function does not, causing loads to fail.  Both now explicitly add an extension.