Integrated the report into the experiments; corrected two offset bugs.
The experiment now writes a report on convergence of the optimisation.
This required that the experiment also be able to calculate the correct
target size for the output, so that this can be used by the report to
generate output images.

In the process of integrating this, it became obvious that there were
a number of issues with the way in which the experiment passed output
to the results (as used by the report).  Specifically:

  - the aggregated patches for volumes were not properly cleared on
    reconstruction of a volume, leading to the set of these growing
    over time (albeit with some patches having been cleared).  The
    correct subset of patches is now removed.

  - the results for the previous epoch were not cleared, leading to
    large amounts of data aggregating on disk; this was a conscious
    choice during development, but not tenable for real experiments
    given the amount of data to be generated.  Data for epochs prior
    to the current is now deleted on each new epoch.