Added a better intitialisation method for convolutional layers.
The previous method for initialising the weights at a convolutional
layer used a 0-mean unit gaussian, which resulted in both a large
number of 0 weights, but also the presence of a reasonable number
of weights clsoe to 1.  Both of these are problematic:

  - 0 weights are impede the back propagation of the error, as the
    gradient of the error w.r.t. the input to a layer is generally
    a multiple of the weights, which in this case is 0.

  - weights close to 1 are problematic as such large weights will
    lead to ever increasing outputs, which can lead to explosively
    large gradients, and then oscillation and thus to instability
    of the network during optimisation.

We now use a method which is close to, though not exactly the same
as, Xavier initialisation, in which the initialisation is drawn
from a uniform distribution on a range which decreases with the
increase in inputs and outputs.  Our problem here is that dense
inference means the number of inputs and outputs is variable, so
instead we use the number of weights in the layer as a heuristic
(as this would be the number of inputs if the patch size was fixed
to the input field of the kernel).  While not theoretically correct
this yields acceptable results in practice.