Corrected (another) issue in the calculation of softmax.
The previous reformulation of softmax intended to prevent underflow
or overflow was, as it turns out, not quite right: I had pushed a log
through a sum, as if it was a product.  This was obviously incorrect.
This has been fixed, and the usual log sum exp trick, of taking the
maximum exponent in a log of sums, subtracting this from all of the
other exponents and adding it to the log, has been used instead.  The
output actually sums to one now, which is nice.