Corrected the calculation of the batch size and number of batches.
The number of batches per iteration was calculated by dividing the
target region size by the patch size, but this is incorrect:  the
number of patches is actually ((volume - patch) / stride) + 1, as
input patches can, and in practice do overlap (output patches do
not, but the calculation from the output size is different again).

This error arose in part because this calculation was performed in
several different places, and while this bug had been corrected in
most of them, it had hidden here in the iteration over batches.  I
have for this reason added most of the required logic to compute the
number of patches in various iterations to the sample parameters.
It is not really a parameter, but it is a simple function of them,
so this seems a reasonable place for it to go.  I have replaced the
other uses with this.

The second issue resolved in this commit was the incorrect calculation
of the batch size for sequential batches.  This arose because of some
complexity around the handling of the end position, as this was used
in multiple contexts some of which required it to be inclusive and
some of which required it to be exclusive.  The current solution is
still not completely correct, as it could allow for an empty subset
of patches, but I will deal with that in the next commit.