Fixed remaining batch offset issue where batch falls on volume boundary.
The previous commit fixed most of the problems with the batch start and
end position calculation, but did not address the problem which would
arise when the batch after the current one began exactly on a volume
boundary.  The previous code would include this volume in the batch, but
generate an empty list of patches for it, which would confuse the patch
extraction code.  This commit now explicitly checks for the two cases in
which the final patch in the current batch should differ from the first
patch in the next batch: where the next batch starts on a new volume, and
where there is no next batch (i.e. we are at the end of the epoch).  In
both cases, end is taken to be one past the last patch in the final volume.