Modified the test / validation / training split to keep the test set invariant.
The testing / validation / training split algorithm is iterative: it takes all
volumes for a given subject and attempts to allocate them to the first subset
in which they will fit.  For this reason the order in which subsets are queried
is significant: changing the size of any subset will change the set of volumes
presented to the subsequent sets, and will thus in most cases change the actual
composition of those sets, even if the size of those sets was not changed.

The original ordering placed the training set first, followed by the validation
set and finally the test set.  The problem with this is that changing the size
of the training set, which is a practical necessity, will thus change the
composition of the test set, which invalidates its use for final results (as
at least some of the volumes in it are likely to have been used during training
or validation at some point during the project.

This commit changes the order of these sets, so that the test set can remain
invariant to changes in the size of the other two sets.  It is obviously not
invariant to changes in its own size, but so long as we stick to one size for
all experiments, the composition of the test set will remain fixed.