Changed the segmentation overlay to show background; flipped png images.
The first change ensures that voxels classed as true negatives for a
binary segmentation overlay show the underlying pixel value; this was
previously masked by the fact that these pixels had a value of 0 for
all RGB components in the segmentation overlay, and hence a value of
0 in the product of that overlay and the underlying image.  This commit
changes the sampling to specify that such voxels now take a value of 1
for all RGB components, and so show as greyscale values in the final
image as expected.

The second change simply flips the png images in Y, as pngs grow down
in Y while the underlying images grow upwards.