Added more detail to the optimisation output (including timings).
The optimiser now includes timings for most of its high level operations,
e.g. graph construction, each batch, each epoch and the full session.

The epoch number is now included in the output for each batch as this is
not always obvious where there is a large number of batches per epoch.

The number of batches in the epoch is also now displayed.