Added the ability to generate a report from existing serialised results.
In the even that a training session does not complete, it is helpful to
be able to generate a report for it retrospectively.  The changes to the
report module required for this are minimal, and for the most part this
commit adds new functionality to the results module, allowing results to
be recomputed from previously saved output.  Given such a restored results
object, writing out the report proceeds in largely the same way as it
would during training.