Modified experiment.run to accept a random seed instead of a generator.
In practice we will never reuse generators between experiments, and any
determinism which we require can be acheived simply by specifying the
same seed.  I have updated the experiment to take just the seed, and
construct the random number generator internally.  This reduces the
amount of code required in each of the concrete experiment modules.