Corrected a numerical stability issue with the categorical cross entropy.
The cross entropy of a distribution q relative to a true distribution p is

    ce( p, q ) = -1  sum_i( p_i log( q_i ) )

which is undefined for q_i == 0 or close enough to this that their floating
point representation is 0.  This commit adds a check which clamps q_i to
the interval [2^{-126}, 1.0] where 2^{-126} is the smallest positive value
representable as a 32 bit floating point number.