Fixed bug in the volume label extraction for the report.
This previously only used the offsets for the representative volumes
for statistics in the dice_for_class list.  While this will normally
include all the volumes selected for overall dice statistics, it need
not, and of course in the case which prompted the change did not.  It
now loads distributions and offsets for the selected volume for all
statistics.