Added a weighted version of the categorical cross entropy cost function.
The intention behind this cost is to assign a higher cost to errors at
voxels for the less probable classes, in order to offset the diminished
influence which these classes exert.  In particular it is hoped that this
might prevent the network from adopting a strategy in which it simply
returns the most probable class in the dataset for every voxel.