Corrected a bug in the order of the test / validation / training split.
The order in which the algorithm fills these subsets was changed in a
previous commit, but this was not reflected in the order of the arguments
to the dataset constructor.  While it is not necessary that these match,
I had already begun to think of them in that order, and so passed the
arguments in that way.  Given this I thought it was better if the order
in the constructors matched the logical order of the subsets.