Added the experiment definition to the report for that experiment.
The experiment definition class which specifies the set up of each
experiment is a comprehensive, and fairly readable summary of the
architecture, model, costs, learning rate, optimiser and dataset
which determine the experiment.  Instead of inventing a descriptor
to record these details, I decided to standardise the format (as
an ExperimentDefinition object, and simply save the source along
with the report.

In addition, I have created a small wrapper script which takes the
name of an experiment module, and provided that this defines an
ExperimentDefinition class, it will do the work of setting up the
experiment and either running this, or generating a report for it
from the results saved on a previous run.