Added the initial set of four basic CNN experiments.
These experiments evaluate the effect of kernel size (either 3^3 or 5^3)
and weighting on the cross entropy cost (all four combinations).  They
run for only a single epoch, but with the full 250 volume training set
(1140 batches of 1000 patches, with 10^3 labels per patch, or roughly
1.14 billion training samples).  This should be enough to get a sense
of their performance.

Future experiments on these architectures should examine the learning
rate and the regularisation parameters.  As the batch only uses about
4GB on the GPU it might also be worth upping the batch size (for smaller
networks such as these) or upping the network size (if these experiments
do not indicate overfitting).