Added support for running multiple experiments via experiment_run.
In order to most efficiently use system resources, it is useful to
be able to queue long running experiments to run one after the other
as this avoids wasted time between experiments.  While this could
be done from the shell, this is a fragile approach, as failure of
an experiment can easily bring down a loop (unless run using &, but
this would effectively run all jobs together, which would degrade
performance given the limited resources of the GPU).

This commit adds support for passing multiple experiments to the
experiment_run module, and to make the syntax for this a little nicer
it also adds proper argument handling via optparse.  The experiments
are run within a try statement, so that should one fail, the others
will still be run.