Added support for different initialisation methods in the convolutional layer.
The previous initialisation was based on the Glorot method, but used an incorrect
estimate of the fan-in (the number of input values to the layer).  I have now
corrected this, and in the process also added two variants: the first is to use
either a uniform or a normal distribution (prevoiusly this was uniform only) and
the second is to orthogonalise the weights.  Neither of these has been evaluated
as yet, but it is necessary to commit them first before evaluation of another
machine.