Renamed experiment class, added automatic inference of experiment name.
The existence of both a module called experiments and a directory with the
same name was causing random import failures, so I have changed the module
name to be different (it is no longer plural).

The ExperimentDefinition class can now determine the name of the module
which defines the concrete definition subclass, and this will suffice for
the experiment name in most cases (it is what was always used in practice).
It is still possible to override this in the subclass if that is really
required, but for the most part this should help to make creating an
experiment definition less fiddly: I had several times copied definitions
but forgotten to update the experiment name, only to have it overwrite the
results of ithe original experiment from which the new one had been copied.
This change prevents at least this error.

The existing experiments have been updated to reflect this change.