Added support for continuing training from an existing model.
This commit completes the deserialisation support in the model: this
was partially supported before, but required that the deserialised
parameters had been put into the correct shape and order.  Doing this
requires that layers are able to specify the order of their parameters
and so all layers have been updated to do this.

The experiment class now takes an initial_epoch, which defaults to 0,
which, if positive, will be used to load parameters for a previously
trained model (at the initial epoch - 1) along with the costs and
times for training it.  These parameters are then passed to the
constructor for the model passed to the optimiser.

This change required that the experiment definition no longer creates
the model itself (as this needs to happen after the parameters have
been read) but rather just the architecture for the model.  As this
actually makes experiment definitions simpler, this turns out to be
a nice change with no loss of generality, as the only variation in
the call to the model constructor was the architecture which was
passed.

The final changes were to: the optimiser (which now takes an initial
epoch argument as well); the experiment runner, which now passes
the epoch option, previously only used for the report, to the
constructor for the experiment; the results archive, which now has
a method to simplify reading a single array from an archive (used
for loading the existing costs and times for previous epochs).