Corrected a serious bug in the RMSProp learning rate optimiser.
The rate returned from RMSProp should be a eta / sqrt( Gamma ) where eta is the base rate
and Gamma is the sum of the squared gradients over a window (determined by an attrition
rate).  This formulation ensures that where the graidents are high enough that the sum of
their squares is greater than 1.0, the rate will be reduced reduced, while conversely, if
the gradients are low enough that the sum of their squares is less than 1.0, then the rate
will increase.  Up until this commit however, the code had used a rate eta * sqrt( Gamma )
which of course does the opposite: large gardients generating a Gamma > 1 will increase
the rate, leading to instability, while small gradients leading to a Gamma < 1 will cause
the rate to shrink.  This resulted in the network getting stuck at local minima in flat
regions of the parameter space.  This commit changes the calculation to use the correct
equation.