Corrected an error in the test data in the sample tests for the accessor.
The accessor tests assumed that that the number of patches in a volume
equalled the number of non-overlapping tiles of the patch size in that
vollume; this is incorrect.  The actual number of patches is determined
by both the patch size and the stride.  The bug in the accessor itself
corrected some time ago, but the tests had fallen behind somewhat in the
rush to ge things working.  Fixed now.