Updated the optimisation tests to cover the new monitor interface.
The interface between the optimiser and monitor classes was changed
some time ago, which broke several tests, but there was no time to
fix them at that point.  As of this commit they pass again (and are
somewhat more complete as well, as the old tests did not include an
assertion based test for parameter updates).