Corrected the check for the epoch parameter in experiment_run.
This check previously treated the epoch parameter as a boolean, and
interpretted False to mean that the parameter had not been supplied.
This is incorrect, as 0 is a valid value to pass, and yet will cause
the parameter to evaluate to False when used as a boolean.  Testing
for None is the proper way to do this.